Phase 1.2: test harness, window arithmetic, CI
The harness lands before the rewrite because Phase 5's leakage gates ARE
tests; written later they become one-off scripts that stop being run.

- app/windows.py: the 60s scoring grid (ADR-001/002). Half-open windows
  make the past-only rule arithmetic rather than convention.
- tests/property: 8 hypothesis invariants -- additivity, past-only,
  alignment idempotence, covering/contains agreement. These guard bugs
  that corrupt every downstream number without ever raising.
- tests/unit: the Phase 0 matrix gate, moved into pytest so it cannot rot.
- CI runs pytest at 10x hypothesis examples; ruff now covers app/ and
  tests/, with v1 modules excluded per-phase until their rewrite lands.

+++ app/core/metrics.py
@@ -45,7 +45,7 @@ class MetricsRegistry:
             series = sorted(self._histograms.get(key, []))
         if not series:
             return None
-        idx = min(len(series) - 1, max(0, int(round((pct / 100.0) * (len(series) - 1)))))
+        idx = min(len(series) - 1, max(0, round((pct / 100.0) * (len(series) - 1))))
         return series[idx]
 
     def snapshot(self) -> dict:
